Update the build the docs documentation
This change updates the instructions that show how to build the
documentation. Technically, at the moment, the Docker container
instructions won't work, as there's some issue with the container that
I'm trying to resolve. But the version that references the locally
installed tools does work. Given that, I'm leaving it for the moment and
will make necessary changes to the relevant files just as soon as I know
more.